Refactor useBindFallbackHandler
