+++ src/cow-react/modules/twapOrders/contrainers/TwapWidget/index.tsx
@@ -12,37 +12,9 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import { formatInputAmount } from '@cow/utils/amountFormat'
 import { WRAPPED_NATIVE_CURRENCY as WETH } from 'constants/tokens'
 import AddressInputPanel from '@src/components/AddressInputPanel'
-import { _getClientOrThrow } from '@cow/api/gnosisSafe'
-import { useWeb3React } from '@web3-react/core'
-import Safe from '@safe-global/safe-core-sdk'
-import EthersAdapter from '@safe-global/safe-ethers-lib'
-// eslint-disable-next-line no-restricted-imports
-import { ethers } from 'ethers'
-import { JsonRpcFetchFunc, Web3Provider } from '@ethersproject/providers'
-import SafeServiceClient from '@safe-global/safe-service-client'
-import { SafeTransaction } from '@safe-global/safe-core-sdk-types'
-
-async function proposeTransaction(
-  safe: Safe,
-  safeService: SafeServiceClient,
-  tx: SafeTransaction,
-  signer: ethers.Signer
-) {
-  const safeTxHash = await safe.getTransactionHash(tx)
-  const senderSignature = await safe.signTransactionHash(safeTxHash)
-  await safeService.proposeTransaction({
-    safeAddress: safe.getAddress(),
-    safeTransactionData: tx.data,
-    safeTxHash,
-    senderAddress: await signer.getAddress(),
-    senderSignature: senderSignature.data,
-  })
-
-  console.log(`Submitted Transaction hash: ${safeTxHash}`)
-}
+import { useBindFallbackHandler } from '@cow/modules/twapOrders/hooks/useBindFallbackHandler'
 
 export function TwapWidget() {
-  const handlerAddress = '0x87b52ed635df746ca29651581b4d87517aaa9a9f'
   const chainId = 100
   const currenciesLoadingInProgress = false
   const allowsOffchainSigning = false
@@ -54,8 +26,6 @@ export function TwapWidget() {
       discount: 0,
     },
   }
-
-  const { account, provider: library } = useWeb3React()
 
   const [frequency, setFrequency] = useState<string>('')
   const [deadline, setDeadline] = useState<string>('')
@@ -117,22 +87,7 @@ export function TwapWidget() {
     console.log('createOrder')
   }
 
-  const bindTwapHandler = useCallback(async () => {
-    if (!library || !account) return
-
-    const client = _getClientOrThrow(chainId, library)
-    const signer = library.getSigner()
-
-    const provider = new Web3Provider(library.send.bind(library) as JsonRpcFetchFunc)
-    const ethAdapter = new EthersAdapter({
-      ethers,
-      signerOrProvider: provider.getSigner(0),
-    })
-    const safe = await Safe.create({ ethAdapter: ethAdapter as any, safeAddress: account })
-
-    const safeTransaction = await safe.createEnableFallbackHandlerTx(handlerAddress)
-    await proposeTransaction(safe, client as any, safeTransaction, signer)
-  }, [library, account])
+  const bindTwapHandler = useBindFallbackHandler()
 
   return (
     <>
